feat: Add support for high and low scan rates in Client class
This commit adds support for high and low scan rates in the `Client` class of the `client.py` file. The `Client` class now accepts two additional parameters, `high_scan_rate` and `low_scan_rate`, which determine the scan rates for high and low scans respectively. The `OnRecv` method has been updated to handle the new scan rates when the user selects the "s" option. After performing a high scan, the program will sleep for `high_scan_second + 3` seconds and then perform a low scan for `low_scan_second + 3` seconds.

The purpose of this change is to provide more flexibility in scanning rates for the `Client` class.